pyproject.toml
[project]
name = "allora_sdk"
version = "0.1.0"
description = "Allora Network SDK"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "spooktheducks", email = "spook@alloralabs.xyz" },
    { name = "gh-allora", email = "gregory@alloralabs.xyz" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "pydantic ~= 2.10",
    "aiohttp ~= 3.12",
    "cosmpy >= 0.11.1",
    "websockets >= 15.0.1",
    "grpclib >= 0.4.8",
    "betterproto == 2.0.0b7",
    "requests ~= 2.32",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-timeout>=2.4.0",
    "mypy==1.11.2",
    "types-cachetools>=5.5.0.20240820",
    "types-chardet>=5.0.4.6",
    "fastapi",
    "starlette",
    "httpx",
]
codegen = [
    "grpcio-tools",
    "betterproto2-compiler",
    "openapi-python-client",
]

[project.urls]
Homepage = "https://github.com/allora-network/allora-sdk-py"
Issues = "https://github.com/allora-network/allora-sdk-py/issues"
"Allora Network" = "https://allora.network"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/allora_sdk"]
include = ["src/allora_sdk/py.typed"]

[tool.hatch.build.targets.sdist]
exclude = [
    "/.github",
    "/.vscode",
    "/.idea",
    "/tests",
    "/scripts",
    "/examples",
    "/.tox",
    "/.venv",
]

[tool.tox]
requires = ["tox>=4.19"]
env_list = ["lint", "type", "3.13", "3.12", "3.11", "3.10", "3.9"]

[tool.tox.env_run_base]
description = "Run tests"
deps = [
    "pytest>=8",
    "pytest-asyncio",
    "pytest-sugar",
    "fastapi",
    "starlette",
    "httpx",
]
commands = [["pytest", { replace = "posargs", default = ["tests"], extend = true }]]

[tool.tox.env.type]
description = "Type check"
deps = ["mypy==1.11.2", "types-cachetools>=5.5.0.20240820", "types-chardet>=5.0.4.6"]
commands = [["mypy", { replace = "posargs", default = ["src", "tests"], extend = true }]]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["asyncio: mark test as async test"]
